requirements.txt
Pillow
split-folders
pdf2image
natsort
opencv-python
pytesseract
transformers[torch]
datasets
seqeval
Pillow
accelerate
huggingface_hub
